[feat] Add new methods to EtudiantDao and InscriptionEtudiantDao
